requirements.txt
amqp==5.0.6
asgiref==3.3.4
autopep8==1.5.7
billiard==3.6.4.0
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.3
django-crum==0.7.9
django-debug-toolbar==3.2.1
gunicorn==20.1.0
html5lib==1.1
idna==2.10
kombu==5.1.0
Pillow==8.2.0
pingsms-api==1.0.3
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
Pyphen==0.10.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1